requirements.txt
aiofiles==0.7.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aioredis==1.3.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
basketball-reference-scraper==1.0.26
bcrypt==4.0.0
beautifulsoup4==4.8.2
blessings==1.7
bs4==0.0.1
cachetools==4.2.2
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
dailymotion==0.2.5
decorator==4.4.2
Deprecated==1.2.13
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.0.12
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.1
future==0.18.2
google-api-core==1.32.0
google-auth==1.35.0
googleapis-common-protos==1.56.4
gpustat==0.6.0
greenlet==1.1.0
grpcio==1.40.0
hiredis==2.0.0
idna==2.8
imageio==2.21.1
imageio-ffmpeg==0.4.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.14.0
lxml==4.4.2
MarkupSafe==2.1.1
moviepy==1.0.3
msgpack==1.0.2
multidict==5.1.0
numpy==1.18.1
nvidia-ml-py3==7.352.0
opencensus==0.7.13
opencensus-context==0.1.2
opencv-python==4.6.0.66
packaging==21.3
pafy==0.5.5
pandas==0.25.3
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
proglog==0.1.9
prometheus-client==0.14.1
protobuf==3.17.3
psutil==5.8.0
py-spy==0.3.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pydub==0.25.1
pyparsing==2.4.7
pyrsistent==0.18.1
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.4.1
ray==1.5.2
redis==3.5.3
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.9
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.4.17
tqdm==4.64.0
typing-extensions==3.10.0.0
unicode==2.9
Unidecode==1.3.4
urllib3==1.25.11
Werkzeug==2.0.0
wrapt==1.14.1
yarl==1.6.3
youtube-dl==2021.6.6
zipp==3.8.1
